refactor(renderer): extract skinning into SkinnedRenderer subsystem
Moves the entire animation/skinning pipeline (rig install, per-frame palette
upload, instanced draw, owning GPU resources) out of NewRenderer.cpp into a
dedicated SkinnedRenderer class.  Graphics team can now iterate on the
skinning side (pipeline + shaders + draw call) without touching NewRenderer
plumbing, and gameplay code talks to it through `renderer.skinned()`.

API change: skinning methods migrate from NewRenderer to SkinnedRenderer.
  before:  renderer.setSkinnedRig(meshes, numJoints);
           renderer.setSkinnedFrame(palette, instances);
  after:   renderer.skinned().setRig(meshes, numJoints);
           renderer.skinned().setFrame(palette, instances);

SkinnedRenderer owns:
  - per-mesh VB / boneVB / IB (set up in setRig, GPU upload is real)
  - palette + instance SSBOs + transfer buffers (grown on demand)
  - skinned pipeline pointer (graphics team to create)
  - CPU-side per-frame staging (palette + instances vectors)

NewRenderer drives the lifecycle:
  - init():     skinnedRenderer_.init(device_)
  - drawFrame:  uploadFrame() before render pass, draw() inside geometry pass
  - quit():     skinnedRenderer_.shutdown() before device teardown

Data types (BoneInfluence, SkinnedInstance, RigMeshSource) stay in
RendererTypes.hpp — they're pure value types and game code constructs them
to call into SkinnedRenderer.

No behaviour change.  Build verified.
